Combined Build (not done yet)
Attached the packets to the standalone build so that we don't have the
assets of the rouge game tutorial.
What mainly needs to be done is to get the touch control functioning and
NPC interactive.
- For the touch control, need to move from keyboard to touch.
- For the NPC interaction, it's probably a problem with collision.
Also added Dimas tile set. I'll have to remap the item icons so that
Unity can cut it.
Keyboard control as of now is WASD or arrow keys. Attack command is on
x, but it doesn't have a function aside from animating.